+++ Domain/src/main/scala/peal/domain/Rule.scala
@@ -1,5 +1,5 @@
 package peal.domain
 
-class Rule(val q: Predicate, val score: Double) {
+class Rule(val q: Predicate, val score: BigDecimal) {
   override def toString = "(" + q.name + " " + "%.2f".format(score) + ")"
 }
